# Pepper some inlines about