This PR adds some `#[inline]`s on a bunch of chosen functions to encourage cross-crate optimizations. In testing I've noticed an improvement in our JSON benchmark:

```
before: test twitter_sval                   ... bench:     332,306 ns/iter (+/- 7,551)
after:  test twitter_sval                   ... bench:     270,247 ns/iter (+/- 7,618)
```

Along with some improvement to a protobuf encoder I'm working on:

```
before: test tests::opentelemetry::buffer_sval  ... bench:         805 ns/iter (+/- 21)
after:  test tests::opentelemetry::buffer_sval  ... bench:         669 ns/iter (+/- 12)
```